+++ src/test/java/domain/game/WinningTest.java
@@ -2,7 +2,6 @@ package domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.card.Winning;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
